reorganização TAD Cilindro
